requirements.txt
pandas
numpy
openpyxl
openai>=1.0.0
pinecone>=5.0.0
python-dotenv
tiktoken
tqdm

# FastAPI and server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Web search
google-search-results>=2.4.2